feat: add constraints in category service

+++ ewm/src/main/java/ru/practicum/ewm/category/service/CategoryService.java
@@ -10,7 +10,10 @@ import ru.practicum.ewm.category.dto.CategoryRequestDto;
 import ru.practicum.ewm.category.model.Category;
 import ru.practicum.ewm.category.repository.CategoryRepository;
 import ru.practicum.ewm.category.util.CategoryMapper;
+import ru.practicum.ewm.event.model.Event;
+import ru.practicum.ewm.event.repository.EventRepository;
 import ru.practicum.ewm.exception.NotFoundException;
+import ru.practicum.ewm.exception.ValidationException;
 import ru.practicum.ewm.user.model.User;
 import ru.practicum.ewm.user.repository.UserRepository;
 
@@ -24,6 +27,7 @@ import java.util.stream.Collectors;
 public class CategoryService {
     private final CategoryRepository categoryRepository;
     private final UserRepository userRepository;
+    private final EventRepository eventRepository;
 
     public List<CategoryDto> getAllCategories(int from, int size) {
         PageRequest pageRequest = PageRequest.of(from / size, size);
@@ -44,6 +48,10 @@ public class CategoryService {
     public CategoryDto updateCategory(long ownerId, CategoryDto categoryDto) {
        Optional<User> optionalUser = userRepository.findById(ownerId);
        if (optionalUser.isPresent()) {
+           Optional<Category> optionalCategory = categoryRepository.findByName(categoryDto.getName());
+           if (optionalCategory.isPresent()) {
+               throw new ValidationException("Категория с указанным именем уже существует");
+           }
            return CategoryMapper.toDto(categoryRepository.save(CategoryMapper.toCategory(categoryDto)));
        }
         throw new NotFoundException("Указанный пользователь не найден");
@@ -53,6 +61,10 @@ public class CategoryService {
     public CategoryDto addCategory(long ownerId, CategoryRequestDto categoryRequestDto) {
         Optional<User> optionalUser = userRepository.findById(ownerId);
         if (optionalUser.isPresent()) {
+            Optional<Category> optionalCategory = categoryRepository.findByName(categoryRequestDto.getName());
+            if (optionalCategory.isPresent()) {
+                throw new ValidationException("Категория с указанным именем уже существует");
+            }
            return CategoryMapper.toDto(categoryRepository.save(CategoryMapper.requestDtoToCategory(categoryRequestDto)));
         }
         throw new NotFoundException("Указанный пользователь не найден");
@@ -62,6 +74,10 @@ public class CategoryService {
     public void deleteCategory(long ownerId, long catId) {
         Optional<User> optionalUser = userRepository.findById(ownerId);
         if (optionalUser.isPresent()) {
+            List<Event> events = eventRepository.findAllByCategory_Id(catId);
+            if (!events.isEmpty()) {
+                throw new ValidationException("Удаление категорий, которые связаны с событиями запрещено");
+            }
             categoryRepository.deleteById(catId);
         }
         throw new NotFoundException("Указанный пользователь не найден");
